requirements.txt
python-jobspy
gspread
pandas
google-auth
google-auth-oauthlib
google-auth-httplib2
